Refactor CommandResult return values in automation, decision, form, process, and task management use cases to use id.value instead of id. Update tenant repository method signatures to use size_t for offset and limit parameters. Modify JSON utility functions to handle size_t instead of uint. Adjust task center use cases to return id.value in CommandResult. Ensure consistency across all use cases and repositories for id handling and pagination parameters.

+++ application-studio/source/uim/platform/application_studio/domain/repositories/project_templates.d
@@ -14,7 +14,7 @@ mixin(ShowModule!());
 interface ProjectTemplateRepository : ITenantRepository!(ProjectTemplate, ProjectTemplateId) {
 
     size_t countByCategoryCount(TemplateCategory category);
-    ProjectTemplate[] findByCategory(TemplateCategory category, uint offset = 0, uint limit = 100);
+    ProjectTemplate[] findByCategory(TemplateCategory category, size_t offset = 0, size_t limit = 100);
     void removeByCategory(TemplateCategory category);
 
 }
